requirements.txt
python-dotenv
requests
airflow
airflow-operators